requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.24.3
gdown==5.1.0
python-dotenv==1.0.1